+++ code/manchester.h
@@ -1,12 +1,14 @@
 #ifndef MANCHESTER_H
 #define MANCHESTER_H
 
+#include <inttypes.h>
+
 void manchester_init();
 void manchester_send(int t);
 int manchester_read();
 
 // low level, for debug
-int manchester_wait_bit();
-void manchester_send_bit(int b);
+int8_t manchester_wait_bit();
+void manchester_send_bit(int8_t b);
 
 #endif /* MANCHESTER_H */
